Change the login interface
Server, login, and password are closely related pieces of data,
so they are now set in the same dialog.
Also, this way we can easily disable all network-related widgets
as long as we are offline.

The new login dialog does not respect encapsulation and the disabling
of widgets has poor error handling, but both are at least as good as before.