Use Qt's open file dialog with preview widget
Since the preview widget is added to the standard dialog's layout,
it might not be displayed properly on other platforms. There used to
be a good way of dealing with this problem in Qt 3, but it does not
seem that it will be reimplemented in the future. Possible workaround:
always use standard (non-native) dialogs.