Fix passing in a local game
The color of the current player was sometimes determined incorrectly when passing in a local game.
The real problem is that there are multiple locations where "current" move is stored and they are
apparently not all updated correctly, but it is anyway a good idea to determine the current player only once
and pass this value to the functions that need it.